[Http] Move middleware from Routing to Server.
