+++ src/Valkyrja/Http/Server/Middleware/CacheResponseMiddleware.php
@@ -11,7 +11,7 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace Valkyrja\Http\Routing\Middleware;
+namespace Valkyrja\Http\Server\Middleware;
 
 use Override;
 use Throwable;
