+++ src/Valkyrja/Http/Server/Middleware/ThrowableCaught/ViewThrowableCaughtMiddleware.php
@@ -11,7 +11,7 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace Valkyrja\Http\Server\Middleware;
+namespace Valkyrja\Http\Server\Middleware\ThrowableCaught;
 
 use Override;
 use Throwable;
@@ -19,12 +19,20 @@ use Valkyrja\Http\Message\Request\Contract\ServerRequestContract;
 use Valkyrja\Http\Message\Response\Contract\ResponseContract;
 use Valkyrja\Http\Middleware\Contract\ThrowableCaughtMiddlewareContract;
 use Valkyrja\Http\Middleware\Handler\Contract\ThrowableCaughtHandlerContract;
-use Valkyrja\Log\Logger\Contract\LoggerContract;
+use Valkyrja\View\Factory\Contract\ResponseFactoryContract;
+use Valkyrja\View\Factory\ResponseFactory;
 
-class LogThrowableCaughtMiddleware implements ThrowableCaughtMiddlewareContract
+class ViewThrowableCaughtMiddleware implements ThrowableCaughtMiddlewareContract
 {
+    /**
+     * The errors template directory.
+     *
+     * @var string
+     */
+    protected string $errorsTemplateDir = 'errors';
+
     public function __construct(
-        protected LoggerContract $logger,
+        protected ResponseFactoryContract $viewResponseFactory = new ResponseFactory(),
     ) {
     }
 
@@ -38,10 +46,17 @@ class LogThrowableCaughtMiddleware implements ThrowableCaughtMiddlewareContract
         Throwable $throwable,
         ThrowableCaughtHandlerContract $handler
     ): ResponseContract {
-        $url        = $request->getUri()->getPath();
-        $logMessage = "Http Server Error\nUrl: $url";
+        $statusCode = $response->getStatusCode();
 
-        $this->logger->throwable($throwable, $logMessage);
+        $response = $this->viewResponseFactory->createResponseFromView(
+            template: "$this->errorsTemplateDir/" . ((string) $statusCode->value),
+            data: [
+                'exception' => $throwable,
+                'request'   => $request,
+                'response'  => $response,
+            ],
+            statusCode: $statusCode
+        );
 
         return $handler->throwableCaught($request, $response, $throwable);
     }
