+++ tests/Unit/Application/Kernel/ApplicationTest.php
@@ -65,15 +65,15 @@ use Valkyrja\Http\Routing\Data\Data as HttpData;
 use Valkyrja\Http\Routing\Dispatcher\Contract\RouterContract as HttpRoutingRouter;
 use Valkyrja\Http\Routing\Factory\Contract\ResponseFactoryContract as HttpRoutingResponseFactory;
 use Valkyrja\Http\Routing\Matcher\Contract\MatcherContract;
-use Valkyrja\Http\Routing\Middleware\CacheResponseMiddleware;
-use Valkyrja\Http\Routing\Middleware\RouteMatched\RequestStructMiddleware;
-use Valkyrja\Http\Routing\Middleware\RouteMatched\ResponseStructMiddleware;
-use Valkyrja\Http\Routing\Middleware\RouteNotMatched\ViewRouteNotMatchedMiddleware;
 use Valkyrja\Http\Routing\Processor\Contract\ProcessorContract;
 use Valkyrja\Http\Routing\Url\Contract\UrlContract;
 use Valkyrja\Http\Server\Handler\Contract\RequestHandlerContract;
-use Valkyrja\Http\Server\Middleware\LogThrowableCaughtMiddleware;
-use Valkyrja\Http\Server\Middleware\ViewThrowableCaughtMiddleware;
+use Valkyrja\Http\Server\Middleware\CacheResponseMiddleware;
+use Valkyrja\Http\Server\Middleware\RouteMatched\RequestStructMiddleware;
+use Valkyrja\Http\Server\Middleware\RouteMatched\ResponseStructMiddleware;
+use Valkyrja\Http\Server\Middleware\RouteNotMatched\ViewRouteNotMatchedMiddleware;
+use Valkyrja\Http\Server\Middleware\ThrowableCaught\LogThrowableCaughtMiddleware;
+use Valkyrja\Http\Server\Middleware\ThrowableCaught\ViewThrowableCaughtMiddleware;
 use Valkyrja\Reflection\Reflector\Contract\ReflectorContract;
 use Valkyrja\Tests\Unit\Abstract\TestCase;
 use Valkyrja\View\Provider\ComponentProvider as ViewComponentProvider;
