+++ tests/Unit/Http/Server/Middleware/ThrowableCaught/ViewExceptionMiddlewareTest.php
@@ -11,21 +11,21 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace Valkyrja\Tests\Unit\Http\Server\Middleware;
+namespace Valkyrja\Tests\Unit\Http\Server\Middleware\ThrowableCaught;
 
 use Valkyrja\Http\Message\Enum\StatusCode;
 use Valkyrja\Http\Message\Request\ServerRequest;
 use Valkyrja\Http\Message\Response\Response;
 use Valkyrja\Http\Middleware\Handler\ThrowableCaughtHandler;
-use Valkyrja\Http\Server\Middleware\LogThrowableCaughtMiddleware;
-use Valkyrja\Log\Logger\Contract\LoggerContract;
+use Valkyrja\Http\Server\Middleware\ThrowableCaught\ViewThrowableCaughtMiddleware;
 use Valkyrja\Tests\Unit\Abstract\TestCase;
 use Valkyrja\Throwable\Exception\Exception;
+use Valkyrja\View\Factory\ResponseFactory;
 
 /**
- * Class LogExceptionMiddlewareTest.
+ * Class ViewExceptionMiddlewareTest.
  */
-class LogExceptionMiddlewareTest extends TestCase
+class ViewExceptionMiddlewareTest extends TestCase
 {
     public function testException(): void
     {
@@ -33,30 +33,41 @@ class LogExceptionMiddlewareTest extends TestCase
         $request    = new ServerRequest();
         $response   = new Response(statusCode: $statusCode);
         $exception  = new Exception();
-        $url        = $request->getUri()->getPath();
 
-        $logger = $this->createMock(LoggerContract::class);
-        $logger->expects($this->once())
-            ->method('throwable')
+        $args = [
+            'exception' => $exception,
+            'request'   => $request,
+            'response'  => $response,
+        ];
+
+        $templateText = 'Error: 500';
+
+        $viewResponse = Response::create(content: $templateText, statusCode: $statusCode);
+
+        $view = $this->createMock(ResponseFactory::class);
+        $view->expects($this->once())
+            ->method('createResponseFromView')
             ->with(
-                self::equalTo($exception),
-                self::equalTo("Http Server Error\nUrl: $url"),
-            );
+                self::equalTo('errors/500'),
+                self::equalTo($args)
+            )
+            ->willReturn($viewResponse);
 
         $handler = $this->createMock(ThrowableCaughtHandler::class);
         $handler->expects($this->once())
             ->method('throwableCaught')
             ->with(
                 self::equalTo($request),
-                self::equalTo($response),
+                self::equalTo($viewResponse),
                 self::equalTo($exception),
             )
-            ->willReturn($response);
+            ->willReturn($viewResponse);
 
-        $middleware = new LogThrowableCaughtMiddleware(logger: $logger);
+        $middleware = new ViewThrowableCaughtMiddleware(viewResponseFactory: $view);
 
         $response = $middleware->throwableCaught($request, $response, $exception, $handler);
 
+        self::assertSame($templateText, (string) $response->getBody());
         self::assertSame($statusCode, $response->getStatusCode());
     }
 }
